refactor(contract): move public-key views and signing into api/keys.rs and api/sign.rs
