+++ crates/contract/src/api.rs
@@ -2,6 +2,8 @@
 
 pub mod authorization;
 pub mod common;
+pub mod keys;
+pub mod sign;
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
 pub mod test_utils;
